Improved layout
Improved handling when the diagram does not contain nested ports. Allow
routing across multiple levels of hierarchy and more optimal placement
of features.
Implemented a two pass layout algorithm for cases when nested ports
exist. The ports will be placed as if the nested ports did not exist and
the nested ports will be placed in a second pass which performs the
final layout.
Updated documentation with new layout preference and updated description
of limitations